small fixes: remove commented code, relocating types and styles
