+++ publisher/src/pages/Reports.tsx
@@ -511,31 +511,6 @@ const Reports: React.FC = () => {
                             Review and Unpublish
                           </ReportActionsButton>
                         )}
-                      {/* {bulkAction === "publish" && selectedRecords.length > 0 && ( */}
-                      {/*  <ReportActionsButton */}
-                      {/*    buttonColor="green" */}
-                      {/*    onClick={() => { */}
-                      {/*      clearAllSelectedRecords(); */}
-                      {/*      clearBulkAction(); */}
-                      {/*      publishOrUnpublishMultipleReports("PUBLISHED"); */}
-                      {/*    }} */}
-                      {/*  > */}
-                      {/*    Review and Publish */}
-                      {/*  </ReportActionsButton> */}
-                      {/* )} */}
-                      {/* {bulkAction === "unpublish" && */}
-                      {/*  selectedRecords.length > 0 && ( */}
-                      {/*    <ReportActionsButton */}
-                      {/*      buttonColor="orange" */}
-                      {/*      onClick={() => { */}
-                      {/*        clearAllSelectedRecords(); */}
-                      {/*        clearBulkAction(); */}
-                      {/*        publishOrUnpublishMultipleReports("DRAFT"); */}
-                      {/*      }} */}
-                      {/*    > */}
-                      {/*      Unpublish */}
-                      {/*    </ReportActionsButton> */}
-                      {/*  )} */}
                     </>
                   )}
                 </ReportActions>
